Added view all recipe functionality
view_all.php is essentially just a copy of the search_recipe.php file,
except instead of searching for a specific string all recipe entries are
queried. The search vars were left in to prevent errors from being
displayed, however they do nothing in the view_all.php file.

A button was added to search_recipe.html to facilitate the call to the
view_all.php file.

Potential TODO: alter the "view all" ability to a "view favorites" or
similar as the size of the DB grows.