Bug Fix: Changed DB char set to fix render bug
Added a line the dump.sql to change the char set used by the DB from
utf8mb3 to utf8mb4 to allow for characters outside of the BMP to be
stored (and subsequently rendered) correctly.

This affects the diacritic on "sauté" and should future proof things
going forward as well.