Removed some whitespace in dump
The extra space(s) after the delimiter was causing the ingredients and
instructions to "behave strangely" when called up for editing.
By "behave strangely" I mean that they were behaving exactly as
specified, with the whitespace in front of them, however this is not the
desired display format so the whitespace has been removed.