Updating dump.sql with new tables
Added new tables for kitchens, cookbooks, and favorites.

These tables currently have no entries and have not been tested.

Additionally, a new field ("cookbook_id") has been added to the recipes
table to point to the cookbook the recipe belongs in. This can be null.

I am not sure if this is even the correct implementation but it is what
I think will work as of right now.