Commenting out cookbook_id in recipe table
Also removing the references to it in the inserted recipes.

Initially I thought that this was the way to do it, but I believe that
it is not after thinking about it. I'm leaving it in as a comment in
case its discovered to be necessary later. If so, the following things
will need to be updated: the recipes inserted in the dump & all
references to recipes db in the php code.