php script to download json recipe
The recipe id is passed in and queried, then placed into an key/value
array and json encoded for raw export.

Users will have to either download or copy out the data depending on
their browser.
Its really not ideal at all, but without overhauling the server's
permissions to r/w files, this enables "exporting" of a recipe
semi-automatically.

Users will have to manually return to the previous page.

Ideally in the future there will be an in-app way to share recipes
amongst users.