Implement upload_recipe
Users can now upload raw json recipes.

!!!HAS NOT BEEN THOROUGHLY TESTED!!!

There is a known bug where refreshing the page after entering a json
recipe re-sends the recipe to the db, thereby re-entering it, even
though the input field is empty.
(which means the browser is caching the POST I believe but I couldn't
get it to reset with what I tried at this time)